fix(ui/tasks/actions): add update task action for labels changes
