+++ ui/src/organizations/actions/orgView.ts
@@ -13,22 +13,15 @@ import {
 import {client} from 'src/utils/api'
 
 // Actions
-import {
-  Action as TaskLabelsAction,
-  addTaskLabelsFactoryAsync,
-} from 'src/tasks/actions/v2/labels'
-
-// Types
-import {Organization, AppState} from 'src/types/v2'
+import {UpdateTask} from 'src/tasks/actions/v2'
 
 export enum ActionTypes {
   GetTasks = 'GET_TASKS',
   PopulateTasks = 'POPULATE_TASKS',
-  AddTaskLabels = 'ADD_TASK_LABELS',
-  RemoveTaskLabels = 'REMOVE_TASK_LABELS',
+  UpdateTask = 'UPDATE_TASK',
 }
 
-export type Actions = TaskLabelsAction | PopulateTasks
+export type Actions = UpdateTask | PopulateTasks
 
 export interface PopulateTasks {
   type: ActionTypes.PopulateTasks
@@ -61,6 +54,3 @@ export const createTaskFromTemplate = (
     dispatch(notify(importTaskFailed(error)))
   }
 }
-export const addTaskLabelsAsync = addTaskLabelsFactoryAsync(
-  (state: AppState) => state.orgView.tasks
-)
