+++ ui/src/organizations/reducers/orgView.ts
@@ -7,7 +7,6 @@ import {
 } from '@influxdata/influx'
 import {Dashboard} from 'src/types'
 import {Actions, ActionTypes} from 'src/organizations/actions/orgView'
-import {addTaskLabels, removeTaskLabels} from 'src/tasks/reducers/v2/utils'
 
 export interface OrgViewState {
   tasks: Task[]
@@ -31,15 +30,11 @@ export default (state = defaultState, action: Actions): OrgViewState => {
   switch (action.type) {
     case ActionTypes.PopulateTasks:
       return {...state, tasks: action.payload.tasks}
-    case ActionTypes.AddTaskLabels: {
-      const {taskID, labels} = action.payload
+    case ActionTypes.UpdateTask: {
+      const {task} = action.payload
+      const tasks = state.tasks.map(t => (t.id === task.id ? task : t))
 
-      return {...state, tasks: addTaskLabels(state.tasks, taskID, labels)}
-    }
-    case ActionTypes.RemoveTaskLabels: {
-      const {taskID, labels} = action.payload
-
-      return {...state, tasks: removeTaskLabels(state.tasks, taskID, labels)}
+      return {...state, tasks}
     }
     default:
       return state
